Build downloader into library.
RetrieveISH.py contains getISH() which will download an ISH file.
StationLookup.py contains tools to lookup a station. ISH files are stored on the ftp by the stations usaf and wban. Often the icao is more convenient. lookupUSAFWBAN() will attempt to look up a stations usaf and wban, given the icao and a year. The year is required because an icao can be carried over when a station is moved or updated, creating a duplicate icao in NOAA's records.